requirements.txt
pandas
numpy
akshare
matplotlib 
mplfinance
python-telegram-bot
python-dotenv